requirements.txt
albumentations==0.4.3
click==8.1.8
editdistance==0.8.1
einops==0.8.1
imageio==2.35.1
lightning==2.3.3
lpips==0.1.4
matplotlib==3.7.5
nltk==3.9.1
numpy==1.24.4
omegaconf==2.3.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
pandas==2.0.3
Pillow==9.5.0
pytorch-lightning==2.4.0
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.19.3
scikit-learn==1.3.2
scipy==1.9.3
tensorboard==2.14.0
torch==2.4.1
torchaudio==2.4.1
torchmetrics==1.5.2
torchvision==0.19.1
tqdm==4.67.1
